'15.09.03 Update
  속도조절까지 가능

git-svn-id: https://dev.naver.com/svn/eotktjdrkf@80 29f03ed2-db14-0510-ae99-898e00d8f4b7

+++ FrameWork_03/app/src/main/java/teamwarpcbstuido/IESEKI/game/Monster.java
@@ -17,8 +17,8 @@ public class Monster extends SpriteAnimation
 {
 
    protected int m_speed;
-   protected  int m_speedX;
-   protected  int m_speedY;
+   protected  float m_speedX;
+   protected  float m_speedY;
    protected int xDestiny, yDestiny;
    protected float xDir, yDir;
    protected  int width, height;
